created models, config, types directories in src directory
