+++ server/src/types/index.ts
@@ -1,0 +1,3 @@
+export * from "./applications.js";
+export * from "./jobs.js";
+export * from "./users.js";
