+++ server/src/types/users.ts
@@ -1,0 +1,8 @@
+import { Types, Document } from "mongoose";
+
+export interface IUser extends Document {
+  name: string;
+  email: string;
+  password: string;
+  applications: Types.ObjectId[];
+}
